# LibCompress: Speed up CanonicalCode::read_symbol() slow path