@nico 2024-09-09
(I do have a local fix for #25005, but while it simplifies a whole bunch of stuff and arguably fixes a bug, it does cost some performance. If it ever lands, this is what can happen on top of this PR:

```diff
% diff -u deflate-half.cpp deflate-full.cpp                     
--- deflate-half.cpp	2024-09-09 18:07:46
+++ deflate-full.cpp	2024-09-09 18:11:02
@@ -128,7 +128,7 @@
             next_code++;
         }
 
-        u32 sentinel = next_code;
+        u32 sentinel = (uint32_t)(next_code << (15 - code_length));
         code.m_first_symbol_of_length_after.append(sentinel);
         VERIFY(code.m_first_symbol_of_length_after[code_length] == sentinel);
 
@@ -166,15 +166,16 @@
         return symbol_value;
     }
 
-    auto code_bits = TRY(stream.read_bits<u16>(m_max_prefixed_code_length + 1));
-    code_bits = fast_reverse16(code_bits, m_max_prefixed_code_length + 1);
+    auto code_bits = TRY(stream.peek_bits<u16>(15));
+    code_bits = fast_reverse16(code_bits, 15);
 
     for (size_t i = m_max_prefixed_code_length + 1; i <= 15; ++i) {
         if (code_bits < m_first_symbol_of_length_after[i]) {
+            code_bits >>= 15 - i;
             auto symbol_index = (uint16_t)(m_offset_to_first_symbol_index[i] + code_bits);
+            TRY(stream.discard_previously_peeked_bits(i));
             return m_symbol_values[symbol_index];
         }
-        code_bits = code_bits << 1 | TRY(stream.read_bit());
     }
 
     return Error::from_string_literal("Symbol exceeds maximum symbol number");
```

It's a bit less work in the hot loop of the slow path, but since my local fix requires adding a check to the hot loop of the fast path, it's currently a net loss.)