@nico 2024-09-10
(macOS system `unzip` needs 682.4 ms ±  5.7 ms per `hyperfine 'unzip -o ~/Downloads/enwik8.zip'`, so quite a ways to go after this still.)